requirements.txt
black==22.3.0
cffi==1.15.0
click==8.1.3
cryptography==37.0.2
isort==5.10.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
pycparser==2.21
PyYAML==6.0
tomli==2.0.1
typing-extensions==4.2.0
